Rename labor to labour in LSR parameters and variables
Use British English spelling throughout the labour supply response
module: directory names, file names, parameter labels, descriptions,
import paths, and test references. Fixes #853.
