+++ EASY_ISSUES.md
@@ -1,0 +1,12 @@
+# Easy-to-Address Issues
+
+1. **#853** - `labor` -> `labour` in LSR parameters — Simple string rename in parameter files.
+2. **#665** - Remove `name` metadata tag — Remove an unused metadata field from parameter/variable files.
+3. **#451** - Change label for `hours_worked` — Change a single variable's label to "Hours worked per week".
+4. **#634** - Incorrect description on non-residential SDLT rate — Fix a wrong description string.
+5. **#614** - CTC calibration budgetary impact description states WTC — Fix a mislabeled description in a YAML file.
+6. **#625** - Broken image in documentation — Fix a broken image link in the docs.
+7. **#920** - Remove deps redundant with policyengine-core — Remove duplicate dependencies from the package config.
+8. **#1371** - Default `would_claim_uc` to True — Change a single default value from `False` to `True`.
+9. **#414** - Remove allowance variables that are only parameters — Delete simple wrapper variables that just return a parameter value.
+10. **#1015** - Make minimum wage for non-apprentices a scale parameter based on age — Restructure a YAML parameter file to use a scale format.
