+++ changelog.d/853.md
@@ -1,0 +1,1 @@
+- Rename `labor` to `labour` in LSR parameter and variable paths, labels, and descriptions to use British English spelling
