--- policyengine_uk/parameters/gov/simulation/labor_supply_responses/income_elasticity.yaml
@@ -1,6 +1,0 @@
-description: Percent change in labor supply given a 1% change in disposable income. This applies only to employment income.
-values:
-  2020-01-01: 0
-metadata:
-  unit: /1
-  label: Income elasticity of labor supply
