--- policyengine_uk/parameters/gov/simulation/labor_supply_responses/obr_elasticities.yaml
@@ -1,6 +1,0 @@
-description: Use the OBR's progression elasticities for the labor supply responses.
-values:
-  2000-01-01: 0
-metadata:
-  unit: currency-GBP
-  reference: none
