--- policyengine_uk/parameters/gov/simulation/labor_supply_responses/substitution_elasticity.yaml
@@ -1,6 +1,0 @@
-description: Percent change in labor supply given a 1% change in the effective marginal wage. This applies only to employment income.
-values:
-  2020-01-01: 0
-metadata:
-  unit: /1
-  label: Substitution elasticity of labor supply
