+++ policyengine_uk/parameters/gov/simulation/labour_supply_responses/obr_elasticities.yaml
@@ -1,0 +1,6 @@
+description: Use the OBR's progression elasticities for the labour supply responses.
+values:
+  2000-01-01: 0
+metadata:
+  unit: currency-GBP
+  reference: none
