+++ policyengine_uk/parameters/gov/simulation/labour_supply_responses/substitution_elasticity.yaml
@@ -1,0 +1,6 @@
+description: Percent change in labour supply given a 1% change in the effective marginal wage. This applies only to employment income.
+values:
+  2020-01-01: 0
+metadata:
+  unit: /1
+  label: Substitution elasticity of labour supply
