+++ policyengine_uk/tests/test_behavioral_responses.py
@@ -1,5 +1,5 @@
 """
-Tests for behavioral labor supply responses.
+Tests for behavioral labour supply responses.
 
 This test module validates that the behavioral response system works correctly
 and that all the critical fixes are functioning:
@@ -29,7 +29,7 @@ requires_hf_data = pytest.mark.skipif(
 
 # Load YAML test cases
 yaml_file = (
-    Path(__file__).parent / "behavioral_responses" / "test_labor_supply_responses.yaml"
+    Path(__file__).parent / "behavioral_responses" / "test_labour_supply_responses.yaml"
 )
 with open(yaml_file, "r") as f:
     yaml_content = f.read()
@@ -37,7 +37,7 @@ with open(yaml_file, "r") as f:
 
 
 class TestBehavioralResponses:
-    """Test behavioral labor supply responses functionality"""
+    """Test behavioral labour supply responses functionality"""
 
     def test_yaml_file_structure(self):
         """Test that YAML file loads correctly and has expected structure"""
