+++ policyengine_uk/variables/gov/simulation/labour_supply_response/income_elasticity_lsr.py
@@ -1,17 +1,17 @@
 from policyengine_uk.model_api import *
 
 
-class substitution_elasticity_lsr(Variable):
+class income_elasticity_lsr(Variable):
     value_type = float
     entity = Person
-    label = "substitution elasticity of labor supply response"
+    label = "income elasticity of labour supply response"
     unit = GBP
     definition_period = YEAR
     requires_computation_after = "employment_income_behavioral_response"
 
     def formula(person, period, parameters):
-        lsr = parameters(period).gov.simulation.labor_supply_responses
+        lsr = parameters(period).gov.simulation.labour_supply_responses
         employment_income = max_(person("employment_income_before_lsr", period), 0)
-        wage_change = person("relative_wage_change", period)
+        income_change = person("relative_income_change", period)
 
-        return employment_income * wage_change * lsr.substitution_elasticity
+        return employment_income * income_change * lsr.income_elasticity
