+++ policyengine_uk/variables/input/employment_income_before_lsr.py
@@ -4,7 +4,7 @@ from policyengine_uk.model_api import *
 class employment_income_before_lsr(Variable):
     value_type = float
     entity = Person
-    label = "employment income before labor supply responses"
+    label = "employment income before labour supply responses"
     unit = GBP
     definition_period = YEAR
     uprating = "gov.economic_assumptions.indices.obr.average_earnings"
